finish

+++ FRONT_END/src/screens/meeting/components/CriteriaEvaluation.tsx
@@ -1,15 +1,74 @@
 import { Button, Textarea } from "@nextui-org/react";
-import { useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
+import { TJob } from "@/api/jobApi";
+import applicantReportApi, { IDetailCriteria } from "@/api/applicantReportApi";
+import { toast } from "react-toastify";
+import { useRouter } from "next/navigation";
 
-const CriteriaEvaluation: React.FC = (): React.JSX.Element => {
+type CriteriaEvaluationProps = {
+  job: TJob;
+  applicantReportIds: string[];
+};
+
+const CriteriaEvaluation: React.FC<CriteriaEvaluationProps> = ({
+  job,
+  applicantReportIds,
+}): React.JSX.Element => {
+  const router = useRouter();
   const targetRef = useRef<HTMLDivElement | null>(null);
-
+  const [detailsCriteria, setDetailsCriteria] = useState<IDetailCriteria[]>([]);
   const [criteriaIndex, setCriteriaIndex] = useState<number>(0);
+  const [updateApplicantReportSegment, setUpdateApplicantReportSegment] =
+    useState<string>("");
 
-  // Hàm cuộn đến phần tử
+  //update
+  useEffect(() => {
+    const intervalId = setInterval(
+      async () => {
+        const { data, status } = await applicantReportApi.updateApplicantReport(
+          {
+            applicantReportIds,
+            applicantReport: {
+              details: detailsCriteria,
+            },
+          }
+        );
+
+        if (!data?._id) {
+          if (status === 401) {
+            toast.error("Unauthorized. Please login again.");
+            clearInterval(intervalId);
+            router.push("/login");
+            return;
+          }
+          toast.error(`Error updating!`);
+          setUpdateApplicantReportSegment("update-failed");
+        } else {
+          setUpdateApplicantReportSegment("update-successfully");
+        }
+      },
+      updateApplicantReportSegment === "update-failed" ? 10000 : 1000
+    );
+    return () => clearInterval(intervalId);
+  }, [detailsCriteria, updateApplicantReportSegment]);
+
+  useEffect(() => {
+    if (!job) return;
+
+    const { criterias } = job;
+
+    setDetailsCriteria(
+      criterias.map((criteria) => ({
+        criteriaName: criteria.criteriaName,
+        comment: "",
+      }))
+    );
+  }, [job]);
+
   const handleScrollToElement = (index: number) => {
     if (targetRef.current) {
-      const elementHeight = targetRef.current.scrollHeight / 12;
+      const elementHeight =
+        targetRef.current.scrollHeight / job.criterias.length;
       targetRef.current.scrollTo({
         top: elementHeight * index,
         behavior: "smooth",
@@ -18,13 +77,31 @@ const CriteriaEvaluation: React.FC = (): React.JSX.Element => {
 
     setCriteriaIndex(index);
   };
+
+  const handleCommentChange = (index: number, value: string) => {
+    setDetailsCriteria((prevs) =>
+      prevs.map((detailsCriteria, detailsCriteriaIndex) => {
+        if (detailsCriteriaIndex === index) {
+          return {
+            ...detailsCriteria,
+            comment: value,
+          };
+        }
+        return detailsCriteria;
+      })
+    );
+  };
+
+  if (!job) {
+    return <p>Loading...</p>;
+  }
+
   return (
-    <div className="max-h-[75vh] flex flex-col gap-8">
-      <h1>Criteria Evaluation</h1>
-      <div className="bg-themeOrange h-1/4 px-4 py-2 rounded-lg custom-scrollbar-thin">
-        {Array.from({ length: 12 }).map((_, index) => (
+    <div className="max-h-[75vh] flex flex-col gap-8 p-2">
+      <div className="bg-themeOrange h-1/4 px-2 rounded-lg custom-scrollbar-thin">
+        {job?.criterias?.map((criteria, index) => (
           <Button
-            key={index}
+            key={criteria?._id}
             className={`block truncate ${
               criteriaIndex === index
                 ? "bg-themeWhite text-themeDark"
@@ -35,7 +112,7 @@ const CriteriaEvaluation: React.FC = (): React.JSX.Element => {
             onPress={() => handleScrollToElement(index)}
           >
             <p className="text-md font-bold">
-              {index + 1}. Criteria {index}
+              {index + 1}. {criteria.criteriaName}
             </p>
           </Button>
         ))}
@@ -44,18 +121,25 @@ const CriteriaEvaluation: React.FC = (): React.JSX.Element => {
         ref={targetRef}
         className="flex flex-col gap-4 overflow-hidden h-3/4"
       >
-        {Array.from({ length: 12 }).map((_, index) => (
-          <div key={index}>
+        {job?.criterias.map((criteria, index) => (
+          <div key={criteria._id}>
             <p className="text-themeDark text-md font-bold">
-              {index}. Criteria {index}
+              {index + 1}. {criteria.criteriaName}
             </p>
             <div>
               <Textarea
+                onValueChange={(value: string) => {
+                  handleCommentChange(index, value);
+                }}
                 minRows={4}
                 fullWidth={true}
                 variant="bordered"
                 placeholder="Add your comment"
                 className="text-themeDark"
+                classNames={{
+                  inputWrapper:
+                    index === criteriaIndex ? "border-themeOrange" : "",
+                }}
               />
             </div>
           </div>
